Some unorganized garbo
Added gameobjects, which the map keeps the list of. Allows rudimentary "sight" by using the distances between them, which the map also sets. Added PromptOptions, which streamlines allowing various different things to be added to the prompts given to the user.

Also created the player object. It tracks the player character and if they're in a structure or not.