Log event part, differentiator numbers in combat
Can store any tag mid-run with the log event part. These serve functionally as booleans (the log either exists or doesn't). The example here is the FOUGHT_GNOLLS tag, which is used to show that the player fought the gnolls (go figure) in place of using an event completed check. With use of ifs, these can log more specific conditions than merely if an event has occurred or not (for example, if a certain character was in your party during an event)

Also, combat with multiple duplicate creeps (or named characters, theoretically) now has logic to append 2, 3, etc to the names of duplicate combatants to distinguish them from each other.