Attribute EPs+conditions, choices w/ EPs, readme,
Added SetAttribute parts so attributes can be added to characters on the fly. Also added the has attribute condition so they can be checked in events.

EventPart.createBy whatever is now in the event part interface instead of in event, so it can be called outside of strictly within events. This was needed for the next big change...

Choice parts can now have nested event parts rather than just the name of an event to go to.

Significantly expanded the README, including all the new features added in this push.

Also cleaned up the event flow somewhat in light of the insert stack. The stack is now all that is used, the functionality from the queuedEvent tag was merged into the stack's functionality, so we don't need to coordinate both anymore.

Character creation events have been added to Redport, including proof of concept for many of the included changes.